requirements.txt
Flask==2.3.3
PyJWT==2.8.0
pytest==7.4.2
pytest-flask==1.2.0